feat(Amazon Linux): Update functions

+++ funcs.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 GREEN='\033[1;32m'
 RED='\033[1;31m'
 YELLOW='\033[1;33m'
@@ -9,57 +11,86 @@ KIND_VERSION="0.11.1"
 CRICTL_VERSION="v1.24.2"
 CRI_LATEST_VER=$(curl -s https://api.github.com/repos/Mirantis/cri-dockerd/releases/latest|grep tag_name | cut -d '"' -f 4|sed 's/v//g')
 VAGRANT_VERSION="2.2.19"
+AWS_OS=$(cat /etc/os-release | awk -F '=' '/^ID=/ {print $2}' | cut -d '"' -f 2)
 
 function install_ansible() {
   echo -e "${YELLOW}[+] Installing Ansible...${NC}"
-  sudo apt -y remove needrestart
-  sudo apt update -y
-  sudo apt install software-properties-common -y
-  sudo apt-add-repository -y --update ppa:ansible/ansible
-  sudo apt install ansible -y
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum update -y
+      sudo amazon-linux-extras install ansible2 -y
+  else [ "${AWS_OS}" = "ubuntu" ];
+      sudo apt -y remove needrestart
+      sudo apt update -y
+      sudo apt install software-properties-common -y
+      sudo apt-add-repository -y --update ppa:ansible/ansible
+      sudo apt install ansible -y
+  fi
 }
 
 function install_docker() {
   echo -e "${YELLOW}[+] Installing Docker...${NC}"
-  sudo apt-get update --yes
-  sudo apt install -y \
-      apt-transport-https \
-      ca-certificates \
-      curl \
-      software-properties-common
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-  sudo add-apt-repository -y \
-      "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-      $(lsb_release -cs) \
-      stable"
-  sudo apt-get install -y docker-ce
-  sudo usermod -aG docker $USER
-  sudo curl -L "https://github.com/docker/compose/releases/download/$DOCKER_COMPOSE_VERSION/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-  sudo chmod +x /usr/local/bin/docker-compose
-  sudo systemctl start docker
-  sudo systemctl enable docker
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum update -y
+      sudo amazon-linux-extras install docker -y
+      sudo usermod -a -G docker ec2-user
+      id ec2-user
+      newgrp docker
+      wget https://github.com/docker/compose/releases/latest/download/docker-compose-$(uname -s)-$(uname -m) 
+      sudo mv docker-compose-$(uname -s)-$(uname -m) /usr/local/bin/docker-compose
+      sudo chmod -v +x /usr/local/bin/docker-compose
+      # TODO docker-compose is not installed on the instance
+      sudo systemctl enable docker.service
+      sudo systemctl start docker.service
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get update --yes
+      sudo apt install -y \
+          apt-transport-https \
+          ca-certificates \
+          curl \
+          software-properties-common
+      curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+      sudo add-apt-repository -y \
+          "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+          $(lsb_release -cs) \
+          stable"
+      sudo apt-get install -y docker-ce
+      sudo usermod -aG docker $USER
+      sudo curl -L "https://github.com/docker/compose/releases/download/$DOCKER_COMPOSE_VERSION/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+      sudo chmod +x /usr/local/bin/docker-compose
+      sudo systemctl start docker
+      sudo systemctl enable docker
+    else
+        echo "Unsupported OS"
+        exit 1
+  fi
 }
 
 function install_cri-dockerd() {
   echo -e "${YELLOW}[+] Installing CRI-Dockerd...${NC}"
-  sudo apt update
-  sudo apt install -y git wget curl
-
-  if [ $(dpkg --print-architecture) = "amd64" ]; then
+  if [ "${AWS_OS}" = "amzn" ]; then
       wget https://github.com/Mirantis/cri-dockerd/releases/download/v${CRI_LATEST_VER}/cri-dockerd-${CRI_LATEST_VER}.amd64.tgz
       tar xvf cri-dockerd-${CRI_LATEST_VER}.amd64.tgz    
       sudo mv cri-dockerd/cri-dockerd /usr/local/bin/
       rm -rf cri-dockerd-${CRI_LATEST_VER}.amd64.tgz cri-dockerd
-  elif [ $(dpkg --print-architecture) = "arm64" ]; then   
-      wget https://github.com/Mirantis/cri-dockerd/releases/download/v${CRI_LATEST_VER}/cri-dockerd-${CRI_LATEST_VER}.arm64.tgz
-      tar xvf cri-dockerd-${CRI_LATEST_VER}.arm64.tgz
-      sudo mv cri-dockerd/cri-dockerd /usr/local/bin/
-      rm -rf cri-dockerd-${CRI_LATEST_VER}.arm64.tgz cri-dockerd
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt update
+      sudo apt install -y git wget curl
+      if [ $(dpkg --print-architecture) = "amd64" ]; then
+          wget https://github.com/Mirantis/cri-dockerd/releases/download/v${CRI_LATEST_VER}/cri-dockerd-${CRI_LATEST_VER}.amd64.tgz
+          tar xvf cri-dockerd-${CRI_LATEST_VER}.amd64.tgz    
+          sudo mv cri-dockerd/cri-dockerd /usr/local/bin/
+          rm -rf cri-dockerd-${CRI_LATEST_VER}.amd64.tgz cri-dockerd
+      elif [ $(dpkg --print-architecture) = "arm64" ]; then   
+          wget https://github.com/Mirantis/cri-dockerd/releases/download/v${CRI_LATEST_VER}/cri-dockerd-${CRI_LATEST_VER}.arm64.tgz
+          tar xvf cri-dockerd-${CRI_LATEST_VER}.arm64.tgz
+          sudo mv cri-dockerd/cri-dockerd /usr/local/bin/
+          rm -rf cri-dockerd-${CRI_LATEST_VER}.arm64.tgz cri-dockerd
+      else
+          echo "Unsupported architecture"
+      fi
   else
-      echo "Unsupported architecture"
-      exit 1
+      echo "Unsupported OS"
   fi
-
   wget https://raw.githubusercontent.com/Mirantis/cri-dockerd/master/packaging/systemd/cri-docker.service
   wget https://raw.githubusercontent.com/Mirantis/cri-dockerd/master/packaging/systemd/cri-docker.socket
   sudo mv cri-docker.socket cri-docker.service /etc/systemd/system/
@@ -89,30 +120,38 @@ function install_cri-dockerd() {
 
 function install_kubectl() {
   echo -e "${YELLOW}[+] Installing Kubectl...${NC}"
-  sudo apt-get update -y
-  sudo apt-get install -y apt-transport-https ca-certificates curl
-  sudo curl -fsSLo /usr/share/keyrings/kubernetes-archive-keyring.gpg https://packages.cloud.google.com/apt/doc/apt-key.gpg
-  echo "deb [signed-by=/usr/share/keyrings/kubernetes-archive-keyring.gpg] https://apt.kubernetes.io/ kubernetes-xenial main" | sudo tee /etc/apt/sources.list.d/kubernetes.list
-  curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
-  curl -LO "https://dl.k8s.io/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl.sha256"
-  echo "$(cat kubectl.sha256)  kubectl" | sha256sum --check
-  sudo apt-get update -y
-  sudo apt-get install -y kubectl
-  sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
+  if [ "${AWS_OS}" = "amzn" ]; then
+      curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
+      sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get update -y
+      sudo apt-get install -y apt-transport-https ca-certificates curl
+      sudo curl -fsSLo /usr/share/keyrings/kubernetes-archive-keyring.gpg https://packages.cloud.google.com/apt/doc/apt-key.gpg
+      echo "deb [signed-by=/usr/share/keyrings/kubernetes-archive-keyring.gpg] https://apt.kubernetes.io/ kubernetes-xenial main" | sudo tee /etc/apt/sources.list.d/kubernetes.list
+      curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
+      sudo apt-get update -y
+      sudo apt-get install -y kubectl
+      sudo install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
+  else
+      echo "Unsupported OS"
+  fi
 }
 
 function install_helm() {
   echo -e "${YELLOW}[+] Installing Helm...${NC}"
-  curl https://baltocdn.com/helm/signing.asc | sudo apt-key add -
-  sudo apt-get install apt-transport-https --yes
-  echo "deb https://baltocdn.com/helm/stable/debian/ all main" | sudo tee /etc/apt/sources.list.d/helm-stable-debian.list
-  sudo apt-get update --yes
-  sudo apt-get install helm --yes
+  curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
+  chmod 700 get_helm.sh
+  ./get_helm.sh && rm -f get_helm.sh
+  [ ! -e ./helm* ] || sudo rm -f -d ./helm*
 }
 
 function install_terraform() {
   echo -e "${YELLOW}[+] Installing Terraform...${NC}"
-  sudo apt-get install -y unzip
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum install -y unzip
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get install -y unzip
+  fi
   wget https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/terraform_${TERRAFORM_VERSION}_linux_amd64.zip
   unzip terraform_${TERRAFORM_VERSION}_linux_amd64.zip
   sudo mv terraform /usr/local/bin/
@@ -121,7 +160,11 @@ function install_terraform() {
 
 function install_packer() {
   echo -e "${YELLOW}[+] Installing Packer...${NC}"
-  sudo apt-get install -y unzip
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum install -y unzip
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get install -y unzip
+  fi
   wget https://releases.hashicorp.com/packer/${PACKER_VERSION}/packer_${PACKER_VERSION}_linux_amd64.zip
   unzip packer_${PACKER_VERSION}_linux_amd64.zip
   sudo mv packer /usr/local/bin/
@@ -151,10 +194,17 @@ function install_kubens() {
 
 function install_minikube() {
   echo -e "${YELLOW}[+] Installing Minikube...${NC}"
-  sudo apt-get install -y conntrack
-  sudo curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube_latest_amd64.deb
-  sudo dpkg -i minikube_latest_amd64.deb
-  rm -f minikube_latest_amd64.deb
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum install -y conntrack
+      curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
+      sudo install minikube-linux-amd64 /usr/local/bin/minikube
+      rm -f minikube-linux-amd64
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get install -y conntrack
+      sudo curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube_latest_amd64.deb
+      sudo dpkg -i minikube_latest_amd64.deb
+      rm -f minikube_latest_amd64.deb
+  fi
 }
 
 function install_crictl() {
@@ -166,22 +216,35 @@ function install_crictl() {
 
 function install_vagrant() {
   echo -e "${YELLOW}[+] Installing Vagrant...${NC}"
-  sudo apt-get install -y virtualbox
-  wget https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb
-  sudo dpkg -i vagrant_${VAGRANT_VERSION}_x86_64.deb
-  rm vagrant_${VAGRANT_VERSION}_x86_64.deb
+  if [ "${AWS_OS}" = "amzn" ]; then
+      sudo yum install -y yum-utils 
+      ## Add HashiCorp repository
+      sudo yum-config-manager --add-repo https://rpm.releases.hashicorp.com/AmazonLinux/hashicorp.repo
+      ## Install HashiCorp vagrant
+      sudo yum -y install vagrant
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get install -y virtualbox
+      wget https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb
+      sudo dpkg -i vagrant_${VAGRANT_VERSION}_x86_64.deb
+      rm vagrant_${VAGRANT_VERSION}_x86_64.deb
+  fi
 }
 
 function install_td-agent() {
   echo -e "${YELLOW}[+] Installing td-agent...${NC}"
-  wget http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
-  sudo dpkg -i libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
-  sudo curl -fsSL https://toolbelt.treasuredata.com/sh/install-ubuntu-focal-td-agent4.sh | sh
-  sudo apt --fix-broken install
-  sudo apt upgrade -y
-  sudo systemctl start td-agent.service
-  [ ! -e ./libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb ] || rm -f ./libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
-  [ ! -e ./td-agent-* ] || rm -f ./td-agent-*
+  if [ "${AWS_OS}" = "amzn" ]; then
+      curl -L https://toolbelt.treasuredata.com/sh/install-amazon2-td-agent4.sh | sh
+  elif [ "${AWS_OS}" = "ubuntu" ]; then
+      sudo apt-get install -y td-agent
+      wget http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
+      sudo dpkg -i libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
+      sudo curl -fsSL https://toolbelt.treasuredata.com/sh/install-ubuntu-focal-td-agent4.sh | sh
+      sudo apt --fix-broken install
+      sudo apt upgrade -y
+      sudo systemctl start td-agent.service
+      [ ! -e ./libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb ] || rm -f ./libssl1.1_1.1.1f-1ubuntu2.16_amd64.deb
+      [ ! -e ./td-agent-* ] || rm -f ./td-agent-*
+  fi
 }
 
 function install_crane() {
